+++ HISTORY.rst
@@ -7,3 +7,9 @@ History
 ---------------------
 
 First release.
+
+1.0.0 (2018-01-19)
+---------------------
+
+Second release. Changes include : Faster way to compute the ned and coverage, total removing of the matching evaluation, adding of a function to check that the class file is coherent.
+An important change is also the removal of the distinction "across" and "within", which for most of the measures did not actually mean anything.